requirements.txt
mysqlclient
django
djangorestframework